Fix unavailable storage causes page fails to load, add cancel to downloader, display download icon for files being downloaded

+++ drive/downloader.py
@@ -4,7 +4,7 @@ from .models import *
 from django.conf import settings
 from requests import Timeout
 from .utils import FileUtils
-import time, requests
+import time, requests, humanize
 
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
@@ -15,6 +15,9 @@ downloader_thread = None
 class DownloaderStatus:
     download_progress = {}
 
+
+class StopException(Exception):
+    pass
 
 class Downloader:
 
@@ -27,16 +30,16 @@ class Downloader:
                             .filter(status=DownloadStatus.downloading.value)\
                             .select_related('file')\
                             .order_by('file__last_modified').first()
-            if download_to_do is None:
+            if not download_to_do:
                 download_to_do = Download.objects\
                                 .filter(status=DownloadStatus.queue.value)\
                                 .select_related('file')\
                                 .order_by('file__last_modified').first()
-
-            if download_to_do is not None:
+            if download_to_do:
                 real_path = os.path.join(get_storage().base_path,download_to_do.file.relative_path)
                 has_error = True
                 force_stop = False
+                detailed_status = None
 
                 try:
                     with open(real_path, 'ab') as file_obj:
@@ -44,56 +47,88 @@ class Downloader:
                         pos = file_obj.tell()
                         if pos:
                             headers['Range'] = f'bytes={pos}-'
+                        response = requests.get(download_to_do.source_url, headers=headers,
+                                                verify=False, stream=True, allow_redirects=True,
+                                                auth=(download_to_do.username, download_to_do.password),
+                                                timeout=(settings.DOWNLOADER_CONNECT_TIMEOUT, settings.DOWNLOADER_READ_TIMEOUT))
+                        if response.status_code < 400:
+                            download_to_do.status = DownloadStatus.downloading.value
+                            download_to_do.save()
+                            receiving_size = int(response.headers.get('Content-Length'))
+                            initial_size = pos
+                            curr_downloaded_size = 0
+                            start_time = datetime.now()
 
-                            response = requests.get(download_to_do.source_url, headers=headers,
-                                                    verify=False, stream=True,
-                                                    auth=(download_to_do.username, download_to_do.password),
-                                                    timeout=(settings.DOWNLOADER_CONNECT_TIMEOUT, settings.DOWNLOADER_READ_TIMEOUT))
-                            if response.status_code < 300:
-                                download_to_do.status = DownloadStatus.downloading.value
-                                download_to_do.save()
-                                receiving_size = int(response.headers.get('Content-Length'))
-                                initial_size = pos
-                                DownloaderStatus.download_progress[download_to_do.id] = (float(pos) / (initial_size + receiving_size)) * 100;
+                            DownloaderStatus.download_progress[download_to_do.id] = (float(pos) / (initial_size + receiving_size)) * 100;
+
+                            try:
                                 for data in response.iter_content(chunk_size=1024):
                                     file_obj.write(data)
-                                    pos += 1024
-                                    DownloaderStatus.download_progress[download_to_do.id] = (float(pos) / (initial_size + receiving_size)) * 100;
+                                    curr_downloaded_size += len(data)
+                                    pos += len(data)
+                                    time_diff = max((datetime.now()-start_time).seconds, 1)
+                                    DownloaderStatus.download_progress[download_to_do.id] = "{:.2f}% ({}/s)".format((float(pos) / (initial_size + receiving_size)) * 100, humanize.naturalsize(int(curr_downloaded_size/time_diff)))
 
-                                    if download_to_do.to_stop:
+                                    if Download.objects.get(id=download_to_do.id).to_stop:
                                         force_stop = True
-                                        break
+                                        raise StopException()
+                            except StopException:
+                                pass
 
-                                DownloaderStatus.download_progress.pop(download_to_do.id)
-                                download_to_do.status = DownloaderStatus.stopped.value if force_stop else DownloadStatus.finished.value
-                    has_error = False
+                            DownloaderStatus.download_progress.pop(download_to_do.id)
+                            download_to_do.status = DownloadStatus.stopped.value if force_stop else DownloadStatus.finished.value
+                            has_error = False
                 except IOError as e:
-                    download_to_do.detailed_status = 'IOError : ' - str(e)
+                    detailed_status = 'IOError : ' + str(e)
                 except Timeout:
-                    download_to_do.detailed_status = 'Request timed out when requesting data from ' + download_to_do.source_url
+                    detailed_status = 'Request timed out when requesting data from ' + download_to_do.source_url
+
+                download_to_do = Download.objects.get(id=download_to_do.id)
+                if detailed_status:
+                    download_to_do.detailed_status = detailed_status
 
                 if has_error:
                     download_to_do.status = DownloadStatus.error.value
-                else:
-                    download_to_do.file.size = os.path.getsize(real_path)
 
                 if download_to_do.to_delete_file:
-                    FileUtils.delete_file_or_dir(real_path)
-                    download_to_do.f.delete()
+                    for i in range(0, 3):
+                        end = False
+                        if os.path.exists(real_path):
+                            try:
+                                FileUtils.delete_file_or_dir(real_path)
+                                end = True
+                            except:
+                                pass
+                        if not end:
+                            break
+                        else:
+                            time.sleep(1)
+                    download_to_do.file.delete()
                 else:
                     download_to_do.save()
             time.sleep(5)
 
     @staticmethod
+    def onstart_cleanup():
+        from .views import get_storage
+        downloads_to_clear = Download.objects.filter(to_delete_file=True).select_related('file').all()
+        for download in downloads_to_clear:
+            FileUtils.delete_file_or_dir(os.path.join(get_storage().base_path, download.file.relative_path))
+            download.file.delete()
+
+    @staticmethod
     def start():
         global downloader_thread
+        Downloader.onstart_cleanup()
+
         downloader_thread = Thread(target=Downloader.downloader_loop)
         downloader_thread.daemon = True
         downloader_thread.start()
 
     @staticmethod
     def interrupt(download):
-        download.to_stop = DownloaderStatus.stopped.value
-        if DownloaderStatus.download_progress.get(download.id, None):
+        if not download.to_stop:
+            download.to_stop = True
             download.to_delete_file = True
-        download.save()
+            download.status = DownloadStatus.stopped.value
+            download.save()
